1. Method to Update Task with name and description 2. Setting elapsed time for a task 3. Time Management within each task cell. Halfway implemented. Need more modification

+++ iWork/Model/Task.swift
@@ -11,7 +11,7 @@ class Task {
     var id: String
     var name: String
     var description: String
-    var elapsedTime: Double
+    var elapsedTime: Int
     var createdAt: Double
     var externalDocLinks: [(url:String, description:String)]?
     var status: TaskStatus
